+++ internal/tilgangsportalapi/get_system_role.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"log"
 	"net/url"
+	"strings"
 )
 
 // GetSystemRole gets information about a specific named role / check if role
@@ -29,6 +30,10 @@ func (client *Client) GetSystemRole(roleName string) (*SystemRole, error) {
 		log.Printf("An error was thrown when unmarshaling response from tilgangsportalen.")
 		return nil, err
 	}
+
+	// always set response for field L2Ident and L3Ident to lowercase as API returns uppercase
+	data.L2Ident = strings.ToLower(data.L2Ident)
+	data.L3Ident = strings.ToLower(data.L3Ident)
 
 	log.Printf("Role with name %s was found.", roleName)